+++ src/SciTE/GUI.h
@@ -153,7 +153,7 @@ public:
 	bool CanCall() const {
 		return wid && fn && ptr;
 	}
-	int Call(unsigned int msg, uptr_t wParam=0, sptr_t lParam=0) {
+	intptr_t Call(unsigned int msg, uptr_t wParam=0, sptr_t lParam=0) {
 		switch (msg) {
 		case SCI_CREATEDOCUMENT:
 		case SCI_CREATELOADER:
@@ -179,10 +179,10 @@ public:
 			throw ScintillaFailure(status);
 		return retVal;
 	}
-	int CallPointer(unsigned int msg, uptr_t wParam, void *s) {
+	intptr_t CallPointer(unsigned int msg, uptr_t wParam, void *s) {
 		return Call(msg, wParam, reinterpret_cast<sptr_t>(s));
 	}
-	int CallString(unsigned int msg, uptr_t wParam, const char *s) {
+	intptr_t CallString(unsigned int msg, uptr_t wParam, const char *s) {
 		return Call(msg, wParam, reinterpret_cast<sptr_t>(s));
 	}
 	sptr_t Send(unsigned int msg, uptr_t wParam=0, sptr_t lParam=0);
